Module weezl::encode

A module for all encoding needs.

Structs

CompressedKey

EncodeState

Encoder

The state for encoding data with an LZW algorithm.

Full

IntoStream

A encoding stream sink.

LsbBuffer

MsbBuffer

Simple

Tree

One tree node for at most each code. To avoid using too much memory we keep nodes with few successors in optimized form. This form doesn’t offer lookup by indexing but instead does a linear search.

Enums

FullKey

Constants

SHORT

Traits

Buffer

Stateful