weezl::encode

Trait Stateful

trait Stateful {
    // Required methods
    fn advance(&mut self, inp: &[u8], out: &mut [u8]) -> BufferResult;
    fn mark_ended(&mut self) -> bool;
    fn restart(&mut self);
    fn reset(&mut self);
}

Required Methods

fn advance(&mut self, inp: &[u8], out: &mut [u8]) -> BufferResult

fn mark_ended(&mut self) -> bool

fn restart(&mut self)

Reset the state tracking if end code has been written.

fn reset(&mut self)

Reset the encoder to the beginning, dropping all buffers etc.