weezl

Module error

Structs

BufferResult

The result of a coding operation on a pair of buffer.

StreamResult

The result of coding into an output stream.

VectorResult

The result of a coding operation into a vector.

Enums

LzwError

The error kind after unsuccessful coding of an LZW stream.

LzwStatus

The status after successful coding of an LZW stream.