weezl::decode

Struct DecodeState

struct DecodeState<CodeBuffer, Constants: CodegenConstants> {
    min_size: u8,
    table: Table,
    buffer: Buffer,
    last: Option<DerivationBase>,
    next_code: u16,
    clear_code: u16,
    end_code: u16,
    has_ended: bool,
    is_tiff: bool,
    implicit_reset: bool,
    code_buffer: CodeBuffer,
    constants: PhantomData<Constants>,
}

Fields
min_size: u8

The original minimum code size.

table: Table

The table of decoded codes.

buffer: Buffer

The buffer of decoded data.

last: Option<DerivationBase>

The link which we are still decoding and its original code.

next_code: u16

The next code entry.

clear_code: u16

Code to reset all tables.

end_code: u16

Code to signal the end of the stream.

has_ended: bool

A stored flag if the end code has already appeared.

is_tiff: bool

If tiff then bumps are a single code sooner.

implicit_reset: bool

Do we allow stream to start without an explicit reset code?

code_buffer: CodeBuffer

The buffer for decoded words.

constants: PhantomData<Constants>
Implementations

impl<C: CodeBuffer, CgC: CodegenConstants> DecodeState<C, CgC>

fn new(min_size: u8) -> Self

fn init_tables(&mut self)

fn reset_tables(&mut self)

impl<C: CodeBuffer, CgC: CodegenConstants> DecodeState<C, CgC>

fn next_symbol(&mut self, inp: &mut &[u8]) -> Option<u16>

fn bump_code_size(&mut self)

fn refill_bits(&mut self, inp: &mut &[u8])